chore: reorganze the code
